# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Music Store API",
        "description": "API for managing music store with artists, albums, and songs",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/api/songs": {
            "get": {
                "tags": [
                    "Songs"
                ],
                "summary": "Get all songs",
                "operationId": "getSongs",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "example": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "example": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of songs retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Hey Jude"
                                                    },
                                                    "duration": {
                                                        "type": "integer",
                                                        "example": 423
                                                    },
                                                    "genre": {
                                                        "type": "string",
                                                        "example": "Rock"
                                                    },
                                                    "album": {
                                                        "type": "object"
                                                    },
                                                    "artists": {
                                                        "type": "array",
                                                        "items": {}
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "pagination": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 50
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 4
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Songs"
                ],
                "summary": "Create a new song",
                "operationId": "createSong",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "artist_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Hello"
                                    },
                                    "artist_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "album_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "duration": {
                                        "type": "integer",
                                        "example": 240
                                    },
                                    "genre": {
                                        "type": "string",
                                        "example": "Rock"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Song created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Song created successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Hello"
                                                },
                                                "duration": {
                                                    "type": "integer",
                                                    "example": 240
                                                },
                                                "genre": {
                                                    "type": "string",
                                                    "example": "Rock"
                                                },
                                                "album": {
                                                    "type": "object"
                                                },
                                                "artists": {
                                                    "type": "array",
                                                    "items": {}
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/songs/{id}": {
            "get": {
                "tags": [
                    "Songs"
                ],
                "summary": "Get song by ID",
                "operationId": "showSong",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Song retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Hey Jude"
                                                },
                                                "duration": {
                                                    "type": "integer",
                                                    "example": 423
                                                },
                                                "genre": {
                                                    "type": "string",
                                                    "example": "Rock"
                                                },
                                                "album": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "title": {
                                                            "type": "string",
                                                            "example": "Abbey Road"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "artists": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "The Beatles"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Song not found"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}